Add multiline value support
Fix multiline value support and try harder to keep spacing in arrays.

In order to achieve that, this code makes use of text_store/text_retrieve
with Shellvars_list when available.

The code for in_sync? in the type is also changed to properly manage
multiline entries.

This change makes the module depends on core v2.1.0