Revert "Fix title patterns to not use unsupported proc (#27)"
At the time of release, we were not aware that there was a bug in Puppet
that caused this fix to not work prior to puppet 4.10.4.

This is a revert that will let us release a properly non-breaking
version and then bump the major version number denoting that the next
release has breaking changes.

This reverts commit 5aa572a6b126d219e97db9f2612f651eaa5462a7.